requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.18.0
webdriver-manager>=4.0.0
